- removed tab bar - removed old search controller - replaced search button
